Refactored branch and moved Role.java to model folder, and updated CreateUserController.java and Role.java to make the constructors public
